+++ extension/CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the "al-translation-cleaner" extension will be documented
 ### Added
 
 * Setting: ATC.whenTranslationNotFound
+* Setting: ATC.translationMethod = ask (adds missing translations and prompts for conflicting existing translations)
 
 ### Fixed